2019 Day 3
This was a cool challange which requires the use of set for it to be
fast enough. I got there pretty quickly, but then I hit a snag with
using union instead of intersection. This took a lot of time.

Another bad decision here was using double variable names for x and y,
this lead to a lot of slow downs as well. Fun challenge overall :)

Score:
      --------Part 1--------   -------Part 2--------
Day       Time   Rank  Score       Time  Rank  Score
  3   00:16:58    374      0   00:20:46   275      0