2019 Day 6
Did this one late, didn't turn my alarm clock on :(

Took me a little while to realize that it was a tree, since we can only
orbit one star at a time. After that it was easy for the first task,
just to loop through all the stars and measure their distance to the
center. After that we just add the total number of links and we have the
answer.

For the second, we just plot the path for you and santa and look for the
first common element.

Score:
      --------Part 1--------   -------Part 2--------
Day       Time   Rank  Score       Time  Rank  Score
  6   04:59:55   7225      0   05:06:23  6073      0