2019 Day 5
I didn't get up early enough to get this one done when it was going on
and had to read for exams then.

I love the intcode machine and so it was fun to extend it, but I did it
in the ugliest way possible. The real answer here is of course to
automatically implement a decoder of the opcodes, but I chose to just
write a path for each possible now for hand. This will not scale as more
instructions are added, but for now it works.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  5       >24h  17587      0       >24h  16129      0