2019 Day 7
This was a really interesting one. I love the use of the Intcode
computer, it makes the challenges so fun!

I had a late night out yesterday, so I didn't rise with the sun. I used
about 50 minutes on both tasks and I'm ok with that. The things that
held me back in the first part was the fact that you had to run a
seperate program for each amplifier. This took a little while to figure
out, but once we had that it was smooth sailing.

The second task was the most interesting. My implementation of the
intcode computer is not very complex, so starting and stopping it would
have been very hard. I have to fix this as I'm sure we will see a lot
more of the intcode computer stuff, but for now I got away with using
multiple threads and queues for input. I normally feel like doing
parallel stuff in python is always a pain, but this time it worked
wonderfully.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score