2019 Day 8
This was a mess, a right mess.

I wake up at 5:45 to make sure I will be able to compete. Then I watch
youtube videos and spaceout until 6:21. Nice :P

Then on part one I get the right idea very quickly, but I forget to
update the minimum amount of zero value each iteration, so all lines
pass the check. This takes me a while, but then I pass that.

Part two was worse, because I set it up alright, but got nonsensical
output. What was wrong was that I had misunderstood and thought that the
lines went back and forth like a snake. They did not. This made
everything a lot easier and I solved it shortly after that.

Not a great score and not a great start, but it was fun regardless :)

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  8   00:47:02   2534      0   01:17:53   2494      0