2019 Day 10
Again I feel asleep, so I started on this one around 06:30. Still would
not have granted me a place on the leaderboard, but it's cool.

I've done this exact problem before, but for project euler:
https://projecteuler.net/problem=351 which is a much larger problem. I
remember there that the trick was to for common gcd, but I blanked on
more than that when solving so early in the morning >_<

For part 1 I spent a lot more time than I'd like to admit with logic
errors and division by zero. For some reason I had the idea that the mod
of the longer vector and the shorter one had to be the same. After that
was removed, things worked out. The bruteforce method I employ is slow
as molasses, but it works.

Part 2 was easy in total, hard in detail. Trigonometry under pressure is
never fun, but I quickly remembered atan2. So the plan was to use atan2
and shift it by PI/2, and sort by that. What I did not remember was
worse. First I forgot that atan2 takes y as it's first parameter. Spent
quite a bit of time on this. Then I spent even more time understanding
that our coordinate system has the y axis flipped, so I flipped the way
the coordinates are read. This made it work. In hindsight, I have no
idea why this worked, I don't think it should work on a general basis?

Anyway it worked in the end :P

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 10   01:00:44   1216      0   01:33:00    623      0