2019 Day 11
Intcode makes a return and that always means fun! I really should have
made a seperate module for my intcode by now or made it stopable and
startable, but I have not. I kind of like the fact that all of my
solutions so far are standalone.

Part 1 and Part 2 are both easy in python, due to defaultdict. I look
forward to trying to implement these in other programming languages.

The description is close to something called langtons ants, going to
look at the intcode and see if it really is :)

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 11   00:16:03    244      0   00:21:24    218      0