2019 Day 12
Today is a sad day, because today, I cheated :(

Part one went just fine, I finished at 282. There was only 20 people who
had gotten both stars, this could be my chance. But I just couldn't
figure this out. I needed to go to work and I really wanted it. But I
couldn't find a way to speed it up.

So eventually 100+ people got both stars and the reddit thread unlocked.
I read a comment which mentioned LCM and that is when I realized the
trick to the challenge, which is that each axis is independent. It would
have been awesome to figure this out on my own, but instead I cheated.

Ended up with a rank of 216. I am not going to do this again, as it
ruined a lot of the fun for me. Important to remember that the main fun
here is the puzzles, not to get on the leaderboard. I lost sight of
that.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 12   00:15:40    282      0   00:47:01    216      0