2019 Day 13
This was such a fun day! I have not cleaned up my code yet and I've left
a lot of stuff in, because I want to keep some of the animation code. I
was also not able to do this on time, because of life things. From here
on out though, I'll try to get them done each morning!

Part 1 was easy, but actually took a little time as I had to figure out
that it was as simple as just counting the blocks. The second part is
where this really shines. I love the inclusion of the intcode here.

Initially I was trying to be smart and calculate the minimum work
required to hit the ball. It kind of worked, but there was some
glitches. I still haven't figured these out and I want to sit down and
do this at a point. It would handle 99% of the cases, but need help on
some.

Next up I thought of maybe just making the pedal follow the ball
perfectly, but this also didn't work. I must just have gotten some logic
wrong here, never really got far. I wrote some code to nicely animate
this though, which was sweet.

In the end the solution was simple. Just change the gameboard in the
input file and run it. Had to hack a little on the Intcode
implementation to make it yield on each output, but other than that it
was quite easy.

This is definitivly a task I want to go back to and clean up and solve
in other ways!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 13   03:34:19   4093      0       >24h   6761      0