2019 Day 14
This one was very fun! I didn't get up to try and get on the
leaderboard, but I would not have gotten there anyway.

Part 1 I solved by doing a backwards calculation of needs. This solves
the challenge of reusing excess materials. It was actually easier than I
thought. I was contemplating using z3, but since it couldn't be floats,
I didn't find an angle straight away.

Part 2 was harder. Initially I tried to just run my solution, but
instead requesting larger and larger amounts of fuel, until the ORE ran
out, but this was way to slow. So I switched it to be iterative, adding
one more fuel requirement to the existing requirements. This made it
manageable to bruteforce, but it spent 1 hour and 9 minutes to produce
the answer. There is definitely a faster way to do this, but for now
this is enough. I will look more at it this evening. Gut feeling now
says that I can just multiply the initial find until I get kind of close
to the limit and then just run it normally after that. We will see!

There is a dependency this time, which is tqdm, to keep estimate how how
long it will take.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 14   03:26:48   1413      0   05:23:54   1538      0