2019 Day 19
This was a very easy day compared to the other days, but I hit a bug in
part 2 that almost drove me crazy. The bug ended up that in the input
routine in the intcode interpreter, I used input.pop() when I should be
using input.pop(0). This caused the input to be read in the reverse way,
so y was x and x was y. This worked for the first task, but didn't work
for the second. Once I fixed that it all just worked.

The strategy for the second task was to follow the bottom edge of the
ray until the opposite corner of the beam is also set. Remember to add
99 and not 100.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 19   04:48:56   2542      0   07:17:19   2013      0