2019 Day 20
Finally back to solving these! I finally have some real free time and
I'm going to spend it catching up on AoC!

This was another very fun problem! It was mostly an exercise in parsing
the input as the rest of the task was a simple graph search. My code
ended up being ugly, but I think it is rather robust.

The second part added an interesting twist, but I realized right away
that it was a simple matter of adding an Z dimension. Rest of solution
is exactly the same. I don't know if there is way to calculate the max
possible Z, but I just set it quite high and tried it out.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 20       >24h   2877      0       >24h   2207      0