2019 Day 21
I love these challenges! This one was rather nice as I'm quite good with
boolean algebra. I actually did this one in 26 minutes, so this would
have been the one time I could have been on the leaderboard. I would
have placed 48 ish for both stars. I regret not getting up earlier
today, but its nice knowing there was atleast one challenge I could
have gotten a leaderboard ranking on.

The first part is rather easy, we simply jump if there is ground
4 away and there is any kind of hole in front of us. This is the best
we can do, since there can be situations where skipping a possible
jump could lead to doom. So we just have to assume that it works and
jump. Logically this then is:

J = (NOT (A AND B AND C)) AND D

For the second part, it's just a copy of the first program, except that
we add the clause that we only jump if we can jump again immidiatly or
if we can make a safe jump from the spot after it. This then becomes:

J = (NOT (A AND B AND C)) AND (D AND (H OR (E AND I)))

I'm not sure if the way I encoded my program is the most compact, but it
got the job done. Very happy about todays performance :)

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 21   05:24:35   1519      0   05:35:29   1184      0