2019 Day 17 Part 2
This was a tough one. First I was lazy and implemented it in z3. But
that took way too long.

So I did something I don't like doing, which is make assumptions about
the output. I guessed that we could find such a route, by looking at the
shortest "normal" way to get through the maze. So I built that up and
printed it out.

Then I first solved this by hand actually. It wasn't that hard, since
I know that the first part has to be part of the the A and so forth.
After having it done this way, I continued on a partial solution I've
made before and made modifications to it. Now it can automatically solve
at least some of the inputs.

Really fun task!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 17   03:50:55   2736      0       >24h   5109      0