2019 Day 22 Part 1
Funny problem, but part2 is super hard. I know it involves reversing the
shuffles, and i've been able to do this in the part2 by using modular
inverse, but I'm not able to do it over all the shuffles needed.

I'm going to look at the answers for part 2, since I don't think I'll be
able to do it on my own and this is something I want to learn.