2019 Day 22 Part 2
Well I was right, it had to with Modular arithmetic. I was able to find
a really good answer in the megasolution thread, that explained
everything step by step. This is the solution you see in the part 2.

There is another way to do the last part, which is to create a 2x2
matrix and create the exponent of that, to get the answer we need.

Very hard task, but very nice to learn something new. I've used modular
arithmatic before, but not to this extent. I have a book on number
theory at home, that I think I might dust off and take a look at. I
enjoy these sort of puzzles so much, that it would be a shame not to
give number theory another chance.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 22   17:56:55   3428      0       >24h   1600      0