2019 Day 24
SOOO CLOOOSEEE! I was so close. I got 136 on part one and then I fumbled
the ball majorly on part 2. I would have gotten on the leaderboard part
2 had I not made such a silly mistake.

Part 1 is a version of Conway's Game of Life. It is not hard to code and
I guess it came down to speed coding in this one. I got rank 136 on this
one, which is my best placing this year.

Part 2 is a very cool spin on Part 1, where you have a recursive board.
I really like the fact that this year has themes, we have seen this
before with the maze planet. I made a crucial mistake though. I forgot
that python range operator when used with a "with" statement shadows the
variable you use "with". I was overwriting x and y by using them as
variables for inner loops. Once I fixed that and the fact that I was
doing world.items() instead of world.values(), I got the right answer.

Not my finest hour on part 2, but I did well on part 1. Good day
overall.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 24   00:12:34    136      0   01:14:46    347      0