2019 Day 16 part 2
Had to look at a hint for this one. I tried alot of different
approaches, some which where fancy, some not so fancy. But it was just
to slow. I suspected the answer was going to involve Fast Fourier
Transform, but the first thing mentioned in the thread was "Summed-area
Table". This one simple tweak in addition to my optimization of just
adding the ones and minuses and also only using the values after the
offset, it was finally fast enough.

I learned about Summed-area tables from this exercise, so that is
great. It still uses quite some time, so I'd be interested to look for
faster solutions in the thread.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 16       >24h   6725      0       >24h   6038      0