2019 Day 18 Part 2
This one I kind of gave up on trying. I bruteforced the first part
earlier and tried a couple of different approaches to the second one,
but nothing was working for me. So I ended up throwing in the towel and
just looking at a solution from the solution thread. I copied that
almost verbatim, but removed dependencies outside the stdlib. This is
therefor not really a clear from me, but rather a learning exercise.

So what did the solution that I copied do? First of all it used bitsets,
meaning that it stored the information about keys collected and such in
the form of integers with bits set or not. This saves on spaces and
makes looksup instant.

Secondly it precompted the distance between all keys and stored the
distance and doors inbetween in a map and used this to do a BFS search
on this. It then had an overview over the shortest path that had gotten
a fullpath and pruned the alternatives that was longer than that.

It also used named tuples in a way I haven't seen in a while. The way
it handled the startpositions, it didn't matter if there was few or
many. Very cool!

I knew about the bitset from implementing sudoku once, but I didn't
think to use it here. I had tried to do a BFS over the solution space,
but without the key to key precomputing and bitset I guess it wasn't
fast enough.

Had to admit defeat here, but I learned a lot in the process of copying
the solution.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
18   07:49:09    820      0       >24h   2824      0