2020 Day 1
Another year, another advent of code!

This is my second year participating and I've been looking forward to
this since last year!

As with last year, the tasks are not that hard and I solved it in 2
minutes. Not a great time, but a top 20 time if the site hadn't crashed.
This sucks, but at least I got my input and was able to just keep
refreshing, some people didn't get their input until most people had
finished part 2.

I fumbled a bit on part 2, copied a file over by mistake and had a
syntax fault that cost me a few seconds which was enough to not get top
100. This might have been my only chance :P

This year I'll try to submit the tasks as I did them first and then
clean them up afterwards.

Score:
      -------Part 1--------   -------Part 2--------
Day       Time  Rank  Score       Time  Rank  Score
  1   00:07:01   127      0   00:07:46   112      0