2020 Day 2
Second day nerves!

I would not have gotten a leaderboard placing today, I was simply to
slow reading the task and understanding it. I would have placed much
better on part 1 if I had not missunderstood and thought that the range
was not inclusive on the upper end. I had to wait 1 minute for that
cooldown to wait.

Second part I also missunderstood, I thought it was the range, but
really it was just the two characters. I like the repurposing of the
same input though!

Overall a nice little snack task, but not something I'll remember for
the ages :)

PS: I ate pav bhaji before solving the task, which was nice.

Score:
      -------Part 1--------   -------Part 2--------
Day       Time  Rank  Score       Time  Rank  Score
  2   00:05:45   954      0   00:08:40   763      0