2020 Day 4
Input validation under stress!

So today was maybe the most practical AoC have been in a long time and I
think a lot of people will learn much from it! Input validation is
something that is crucial to get right when making any kind of
application, but it often overlooked!

For the performance itself, in part 1 I only made one mistake, which was
to not think that the last line needs to be validated as well. This
caused me a one off error and caused me to lose one minute. I think that
I'm seeing a pattern here, and this is something that would have been
caught, if I had been using the test inputs.

For part 2 I wrote up the spec and got it right on the first try, but I
think that my inclination towards strong input validation fucked me
here. I took a very defensive position on what kind of inputs might be
thrown at us and assumed the worse. Looking at the input now, I see that
there is really no malicious input here and that a more simple approach
would have worked fine. I also think that the stress got a bit to me, I
was typing badly and so on. For the next part, I think that I should be
more aggressive and just wing it!

I was quite far off getting on the leaderboard, but that is OK :) I look
forward to making this a bit prettier!

Score:
      -------Part 1--------   -------Part 2--------
Day       Time  Rank  Score       Time  Rank  Score
  4   00:06:13   483      0   00:19:12   376      0