2020 Day 8
YEEEEEEEEEEEEEEEESSSSSSSSSSSSSSSSSSSSS!

Finally we have an instruction set task! I hope this continues!

I didn't wake up on time today, so leaderboards where off limits, but I
timed myself and got 00:05:49.26 on the first part and 00:04:43.44 on
the second part for a total of 00:10:43.7. This would not have been on
the leaderboards, so I'm not to sad, but it would have been a very good
placing, so I'll make sure to get up the other days :)

The task themselves where easy so far, but I really liked part 2, which
is unorthodox. Let's hope that this is a permanent change of direction!
I'll try to refactor my VM so it's ready for future tasks.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  8   02:26:23  14118      0   02:31:06  11216      0