2020 Day 9
Today I placed pretty well on the leaderboard, but the task was not
really fun. I must say that I'm sort of dissapointed with this years
challenges :/ This might change though, if there is some smart way to do
this faster than the very naive way that I've implemented it.

I was really fast today and tried to not look at the puzzle at all only
at the input and the task. I think I would have placed a bit better on
task 1 if I had noticed that it was the previous 25, I originally tried
all. I also didn't skip the preamble, but thank god that I tested on the
test cases. It also threw me for a loop that the last input line was a
blank.

This will be nice to try to cleanup.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
  9   00:04:17    276      0   00:07:16    148      0