2020 Day 11
Today I woke up at 6:01 and just had to just rip my laptop open :P
Hardest part of AoC this year has been getting up in the morning!

I'm a bit dissapointed with todays performance, I was nervous and not
really working at optimal pace. Still I solved the task in my head fast
and was just slow to execute.

Part 1 is just normal cellular automata, but on part 2 I stumbled a bit
on the fact that an empty seat also blocks vision. This took some time.
I guess this is the second time.

Both of my solutions are kind of slow here. I wonder if it's faster to
build a list of indexes for all seats in part 2 than to lookup each
time.

I think this whole thing can be converted to a set of equations and done
faster, but I don't know. I seem to remember something about matrix
convolution or something.

This task was not so novel, but it's not super easy too. Good game
overall.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 11   00:16:06    665      0   00:26:48    647      0