20202 Day 12
Slept horribly today, woke up 3 times during the night, so eventually I
gave up and disabled the alarm for 5:30. Got up at 6:25 ish something
and did the task, spent 00:15:55 on part 1 and 00:09:42 on part 2.

The task today was a bit bland for part 1, but it I liked the twist in
part 2. It was much easier than I thought to make it work and I'm quite
proud of the fact that I defined the left and right turn function so
fast and while so sleepy.

I hit a couple of bumps. One was that I didn't remember that with
Manhattan distance you need to take abs of both x and y.

On part 2 I wasted some time thinking I needed to rotate the point, but
I found the left and right turns quite fast. Don't know how I did it
though, but I did :D

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 12   00:46:57   5644      0   00:56:39   3695      0