2020 Day 13
Woohoooo!

Todays part 2 was the first good puzzle this year. My score is super
bad, but I don't really care about it when the puzzle is this good.

On part 1 I mostly lost time due to reading the instructions, the coding
went smooth.

Part 2 however was a much harder nut to crack. First I tried to be
naive, that didn't really work. Then, being the lazy sod I am, I tried
to use z3 to solve it, but that took to long time too. Then I had the
idea to look at the periods between correct matches for just the first
bus and the other busses. I spent a lot of time on this and eventually
figured out that you get both an offset and a period on when the two
buses align on the requirements.

I thought of some LCM stuff, but that is not applicable here because of
the offsets. I then realized that any answer must lie on an offset + a
multiple of the period. I tried to merge the offsets and such together
with just the way I found the 2 first, but this didn't work.

In the end, I realized that I could begin with finding the offset
and period for 1 matching bus, then using that offset and period, find
one for 2 matching buses and then use that to find 3 and so on. Looking
back it's obvious, but I'm not frustrated I didn't see it earlier. This
is just the sort of brain teaser that I want AoC to be!

I hope this sets the standard for the rest of the challenges, because
this was fun!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 13   00:05:29    311      0   01:21:29   2157      0