2020 Day 16
A better task today! I like that it's input validation again and it's
cool to use the python parse package to make the input parsing easy. I
was 2 minutes off the part 2 leaderboard and I really do think I could
have gotten on it, if I had been a bit better. Maybe next time!

I read the task more carefully today and it was really just some nerves
and poor reading and writing speed that made part one kind of slow. On
part 2 though, I was able to figure out a way to do it fast, but I spent
a bit to much time debugging, I should have just gone for it.

The way I solved part 2 was to make a list over positions where the
field could be valid for each field. Then I iterated through the valid
tickets and if a position didn't satisfy the requirements for a field,
I removed it from the list of positions for that field.

When through all the valid tickets, I loop through these lists and see
which fields only have one valid position. I then assign these fields to
those positions. There are still some fields which can be in multiple
positions, so I loop through the lists again and remove those positions
from all field lists. This process repeats until all fields are valid.

There is possibilities for this to not work, if there are fields which
are interchangeable. My code would hang on this, but I made a guess. The
next thing to ensure would be too look if there are positions that only
one filed can have. After that I guess you would have to do some kind of
search... this is where I usually break out z3 and make it work it's
black magic. :P

I wonder if there is an easy "smart" way to solve for even tricky
inputs. Can't wait to see the reddit implementations!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 16   00:08:48    258      0   00:23:15    145      0