2020 Day 17
Bleh, bad time today!

The task is ok in itself, we have had something sort of like this
before, but I botched this one completely. On part 1 it took me forever
to realize that the board was expanding, that the coordinates around had
to be negative and positive and so on.

I don't think I would have gotten on the leaderboard today, but still,
this was a disgrace. I'll try to do better next time.

My solution today was very slow, and I think that is because I chose to
use nested default dicts and to iterate through the ranges on each run.
I think it will run much faster once I change it up to be a list of
active cells and then just the cells around that cell. Lets see!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 17   00:35:06   1426      0   00:43:08   1397      0