2020 Day 18
Good task, bad rank today!

Todays task was fun! I implemented a simple parser for it and modified
it slightly for part 2. I'm pleased with my time in general with this
don't really know if I could have done it much faster.

Only real bad execution was that on part 1, I was altering a global
variable instead of the input argument to the function. So instead of
operating on a smaller part of the expression, I was operating on the
whole expression each time.

There are faster ways to do this. The first that comes to mind is to
avoid the entire problem of parenthesis by simply applying a regex to
find the smallest expression that doesn't have parenthesis in it and
just evaluating that and replacing it. I entertained using regex for
this in the beginning, but didn't catch this trick until I was done.

One FILTHY way of doing this, that I only saw once I went to the
subreddit is to create your own number type in python and redefine the
operators. For part 1 you replace '*' with '-' and for part 2 you
replace '+' with  '*'. This is so fucking clever!

When I clean this up, I'll keep two solutions one with the regex,
because that is what I could come up with and one with with the python
types.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 18   00:30:32   1523      0   00:43:14   1196      0