2020 Day 19
Good task today!

I solved part one by simply expanding all possibilities and checking
those against the list of messages. Part 2 is where it got hard and I
lost a lot of time trying to continue with my dying method.

Right when I saw the task I thought about using regex, but again put it
away, because I didn't know if it would work. Once I switched over to
creating the regex, the only real speed losses was to figure out how to
use a {} literal in a format string and the fact that I forgot to add it
to the second pattern when I was running >_<

Overall very fun challenge and it will be exciting to see how Stensrud
solves this!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 19   00:28:37    516      0   01:16:27    964      0