2020 Day 20
FINALLY A GOOD CHALLENGE! Today was really nice! I had to go after
solving part 1 and I spent so much time over thinking this. I think I
spent maybe 6 hours total, but this could have been done much faster.

The thing I did wrong was that I didn't realize that each pieces border
can only fit a single piece. Had I known this, it would have been a lot
easier, but I solved it thinking many would match.

There was some turning points. When I read that the border pieces would
not match anything, I realized how easy part 1 is. Just check all edges
and see if they match any other edges. If they don't, we mark the tile
they belong too. When we are done, the tiles who are marked twice are
the corners. I realized this after trying to solve the whole board for 3
hours. I actually realized it in the shower and ran out and coded it up
haha!

I solved the problem of stitching the picture together with z3 first,
and it actually ran, but way to slowly for part 2. It was a good
exercise though, because it would have solved ambiguous puzzles too, not
simple ones like this one where each edge only fits one other piece.

Once I scrapped what I had done on part 1 and started organizing the
code more, it was a slow but steady path to victory. I used numpy here
and boy did it make things easier. I'm going to implement a version
without numpy, but not now, now I'm going to bed!

Finding the sea monster was not very hard, just build a list of
positions where it must match as an offset of x and check if those are
set.

Overall I think the task I was solving in my head was a much more fun
and challenging task than the task we were given. So that is the take
away today:

SOLVE THE TASK YOU ARE GIVEN!

From now on, I'll assume that all inputs are nice and friendly, until
proven otherwise.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 20   03:26:09   3552      0   13:45:26   4073      0