2020 Day 22
Woke up at 06:00 today, had snoozed my alarm 3 times :/

I got PTSD flashbacks when I read the space cards name and thought of
day 22 in 2019. I didn't manage to do part 2 there, but I learned a lot
from it. Sadly this task falls far short of that task. This seems to be
another case of tricky to read instructions, which was bad since I just
woke up.

I replicated the exact output from the examples and this helped me find
something I should have read: That the previously seen rule is only for
rounds in this game. I should really read the tasks very carefully from
now on.

I had to upper my recursion limit, because it was timing out. I think
this is because I didn't loop normal rounds in the function. The games
doesn't go very deep, so I think this is easy to fix, but it was quicker
to just hack the recursion limit.

Overall, not too bad day, but still, nothing learned from solving it. I
hope that when I look at the subreddit that I find out that there is a
brilliant way to solve this.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 22   00:13:48   1801      0   00:55:02   1498      0