Cleanup of 2020 Day 20
This ended up being quite verbose, but it doesn't matter, it now runs
fast and without numpy.

I'm quite proud of the orient function. It is totally unnecessary, as
just trying all the orientations until we get it would be enough, but I
optimized it by hand with truth tables and so on.