2020 Day 23
It's my birthday! And it was an ok task today!

I had to use C++, to get the linked list that I needed. C++ iterators
are lifesavers! I could combine a linked list with a vector of iterators
to get O(1) lookup of any element and the list to do fast inserts.

I choose to simply always shift the list so that the first element is
first, to avoid having to check for wrap around on the end.

My rank today was poor, which mostly is me not wanting to switch to C++.
This is the first task where python rather than the implementation has
been the problem and so I was reluctant to spend the effort to implement
it in C++, since I thought that if python didn't cut it, C++ wouldn't
either, but since the numbers seemed so small I had to give it a try.

This is also a reminder that I need to code more in c++, to keep it
sharp. Rust is good for slow stuff, but C++ is good for hacky stuff.

It ends up being OK fast with C++, I can't really wait to go and see on
the subreddit what the proper solution is.

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 23   00:20:36    523      0   01:28:49    935      0