Cleanup of 2020 Day 24
This turned out super neat! Using complex numbers and axial coordinates
really did the trick. What turned it around for me was a comment on
reddit which linked to a picture showing the connection between a
triangular grid and the hexagon system.

Part 2 still takes a bit more time than I want, but I don't know if
there is anything to be done for that.