2020 Day 25
So long, and thanks for all the fish!

Todays task was quite basic in python due to pow having support for mod.
I didn't go that route immediately though and spent some time looking
with Mathematica and so forth instead of bruteforcing. I have
bruteforcing an attempt in the beginning but didn't bother with pow(). I
then started with Mathematica and went back with pow() and didn't see
that it had solved before later. Weak performance by me, but I want to
implement this in C++, since pow with mod could be cool to implement
again!

Second part is just like the first, have all the stars done.

The tasks have been much worse this year, there is not way around it.
Last task last year was to play a game on a computer written in Intcode,
a custom ISA. This year it was 2 for loops. If it hadn't been for the
challenge with Stensrud this would have been a boring year. I'll go back
and do earlier years and see.

Things I've learned about this year:

- Chinese remainder theorem
- XOR to find missing items
- Hexagons
- The parse python module

I wish next year to be better and hope everyone has a nice Christmas!

Score:
      --------Part 1--------   --------Part 2--------
Day       Time   Rank  Score       Time   Rank  Score
 25   00:14:07    790      0   00:14:14    658      0