Cleanup of 2015 Day 14
Eliminated the use of a loop in part one and removed most external
dependencies. I'm sure the second part could be made neater and to use
less memory, but I'm not really too concerned with that. I want to do
multiple tasks per day to really get up to speed with these tasks.