2015 Day 15
This one was weird, part one took way too long, given how I ended up
solving it. Almost all the time was spent generating the possible ways
to configure 100 across the input. Ended up doing a simple brute force
loop here, but I'm not sure this was really the fastest way to do
things. I'm not really sure how I will solve this one for the future,
refining this is going to be some good practice. Some of the mental math
was also not so easy to do.

I ran into one bug here, which is that the negative totals become 0, but
I found this rather quickly, so it was no big hinder. Important reminder
to always read the task description.

Part two was relatively easy to do and most of the time was spent
figuring out how to do the multiplications correctly. In the future I
should really start using np arrays for stuff like this, as it would
have given me a lot of stuff builtin.

Not pleased with my time here, but it's all good training

Part one: 00:29:36
Part two: 00:02:36