2015 Day 16
This was a clusterfuck! I ended up just guessing for both parts! My
program cut down the possibilities to a few and then for the first part
I just chose the biggest one and for the second I tried two times. I'm
not at all sure I understood this task!

I think my conversion to part two was nice, with converting the know
dict to have lambdas as values. Still, today was not nice, given that I
had to guess! I really did read the description carefully, so it's going
to be very interesting to read why this didn't work?

Part one: 00:20.41
Part two: 00:05:09