Cleanup of 2015 Day 16
ARRRGGGGGGG! My logic was totally correct, but there was a bug. In the
beginning I though we should filter out counts with 0 on them from the
list of facts we know about each Sue, but then later that was false. I
removed the part from the dict comprehension which checks if it's equal
to zero, but didn't do so completely. A "if int(v)" was left, which was
removing the zero facts.

Once removed it gave only one answer immediately. I also reworked it so
that we are checking the properties on the sues instead of looping
through the entire known at once.

It is worth pointing out here that pythons "else" keyword is really
elegant when combined with a for!