2015 Day 19
Well this one was the first interesting in this run!

Part one was done on the buss and it took me a bit longer than I wanted.
I wasted a lot of time on the string slicing and the fact that in the
test inputs only had one character replacements, while real input has
replacements of different lengths. It ended up being OK neat though.

Part two was nice! I started trying to just mutate the string into it,
but that is too slow. Then I had the insight of switching the
replacements around and trying that with a BFS. This is also too slow. I
then had the idea to "Just Wing It" (TM), and do a greedy search based
on just following the shortest answer in the pile. This ended up
working!

It is interesting however, it only works sometimes! If it takes the
wrong turn somewhere, it will get stuck on my input. I don't know why
this happens, but I'm thinking it's related to the iteration order over
the "sets".

I did a quick read on the forums for this task and apparently there are
some really elegant solutions. My next move is going to be to try to
figure out why it only works some times and after that I want to look at
the CYK algorithm for getting a fast worst case performance.

Part one: 00:15:22
Part two: 00:23:41